@@ -25,24 +25,24 @@
   },
   "dependencies": {
     "@actions/core": "1.11.1",
-    "@actions/github": "6.0.0"
+    "@actions/github": "6.0.1"
   },
   "devDependencies": {
-    "@commitlint/cli": "19.6.1",
-    "@commitlint/config-conventional": "19.6.0",
-    "@rotki/eslint-config": "3.6.0",
-    "@tsconfig/node20": "20.1.4",
-    "@types/node": "20.17.13",
+    "@commitlint/cli": "19.8.1",
+    "@commitlint/config-conventional": "19.8.1",
+    "@rotki/eslint-config": "4.1.0",
+    "@tsconfig/node20": "20.1.5",
+    "@types/node": "20.19.0",
     "@vercel/ncc": "0.38.3",
-    "bumpp": "9.10.1",
-    "eslint": "9.18.0",
-    "eslint-plugin-github": "5.1.5",
+    "bumpp": "10.1.1",
+    "eslint": "9.28.0",
+    "eslint-plugin-github": "6.0.0",
     "husky": "9.1.7",
     "js-yaml": "4.1.0",
-    "lint-staged": "15.3.0",
-    "prettier": "3.4.2",
+    "lint-staged": "16.1.0",
+    "prettier": "3.5.3",
     "typescript": "5.7.3",
-    "vitest": "2.1.8"
+    "vitest": "3.2.3"
   },
   "engines": {
     "node": ">=20 <21",
